+++ frontend/src/App.tsx
@@ -5,6 +5,7 @@ import ChatWindow from './components/ChatWindow/ChatWindow'
 import ProtectedRoute from './auth/ProtectedRoute'
 import AdminLayout from './admin/AdminLayout'
 import CustomerManager from './admin/CustomerManager/CustomerManager'
+import ShipmentManager from './admin/ShipmentManager/ShipmentManager'
 import './App.scss'
 
 function ChatLayout() {
@@ -32,6 +33,7 @@ function App() {
       >
         <Route index element={<Navigate to="customers" replace />} />
         <Route path="customers" element={<CustomerManager />} />
+        <Route path="shipments" element={<ShipmentManager />} />
       </Route>
     </Routes>
   )